+++ config/migrate-mongo/index.js
@@ -1,5 +1,6 @@
 const Config = require("../index");
 
+console.log(process.env.NODE_ENV);
 module.exports = {
     mongodb: {
         url: "mongodb://localhost/",